EchoNest Living-VoiceAgent/requirements.txt
pyttsx3
SpeechRecognition
PyAudio
python-socketio
flask-socketio
websocket-client